requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
# For SOCKS proxy support, uncomment:
# pysocks>=1.7.1
